backend/requirements.txt
django==1.9.5
djangorestframework==3.3.3
django-extensions==1.6.1
coverage==4.0.3
pytest==2.9.1
pytest-django==2.9.1
ipdb==0.9.3
factory-boy==2.6.1
djangorestframework-jwt==1.8.0
psycopg2==2.6.1
celery==3.1.23
redis==2.10.5